requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
Pillow==8.0.1
pkg-resources==0.0.0
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
